Continued work getting the visualization framework functional. Added openVisN.Query openVisN.Calculations and openVisN.Components namespaces.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS11]$/openHistorian/Main;C21973

+++ Source/Libraries/openVisN/Library/AllSignals.cs
@@ -49,7 +49,7 @@ namespace openVisN.Library
 
         public MetadataBase MakeSignal()
         {
-            return new MetadataSingle(SignalId,(ulong)PointId,DeviceName,Description);
+            return new MetadataSingle(SignalId,PointId,DeviceName,Description);
         }
     }
